# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRITIQ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jan/{code}:
    get:
      operationId: lookupJan
      tags: [products]
      summary: Look up product info by JAN code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: JAN lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JanLookupResult"

  # ── Admin ──────────────────────────────────────────────────
...
  /admin/criterion-suggestions:
    get:
      operationId: adminListCriterionSuggestions
      tags: [admin]
      summary: List criterion suggestions
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CriterionSuggestionItem"
  /admin/criterion-suggestions/{suggestionId}/review:
    patch:
      operationId: adminReviewCriterionSuggestion
      tags: [admin]
      summary: Approve or reject a criterion suggestion
      parameters:
        - name: suggestionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestionReview"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CriterionSuggestionItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/product-suggestions:
    get:
      operationId: adminListProductSuggestions
      tags: [admin]
      summary: List product registration suggestions
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductSuggestionItem"
  /admin/product-suggestions/{suggestionId}/review:
    patch:
      operationId: adminReviewProductSuggestion
      tags: [admin]
      summary: Approve or reject a product suggestion
      parameters:
        - name: suggestionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestionReview"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSuggestionItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CriterionSuggestionItem:
      type: object
      required: [id, categoryId, name, description, status, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: string
        reason:
          type: ["string", "null"]
        status:
          type: string
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ProductSuggestionItem:
      type: object
      required: [id, categoryId, brand, name, status, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        brand:
          type: string
        name:
          type: string
        modelNumber:
          type: string
        janCode:
          type: ["string", "null"]
        price:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    JanLookupResult:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
        name:
          type: string
        brand:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        lowestPrice:
          type: ["number", "null"]
        highestPrice:
          type: ["number", "null"]
...
    SuggestionReview:
      type: object
      required: [status]
      properties:
        status:
          type: string
        adminNotes:
          type: string