# lib/api-zod/src/generated/types/criterionSuggestionItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CRITIQ API
 * OpenAPI spec version: 0.1.0
 */

export interface CriterionSuggestionItem {
  id: number;
  categoryId: number;
  name: string;
  description: string;
  /** @nullable */
  reason?: string | null;
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
}